progress

+++ src/main/java/graphqldemo/repository/PlayerCharacterRepository.java
@@ -3,5 +3,5 @@ package graphqldemo.repository;
 import graphqldemo.model.PlayerCharacter;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PlayerCharacterRepository extends JpaRepository<PlayerCharacter, Long> {
+public interface PlayerCharacterRepository extends JpaRepository<PlayerCharacter, Integer> {
 }
